[fix] 무한스크롤 오류 수정 새로 컴포넌트 생성  #108
