+++ src/pages/Exhibition.jsx
@@ -1,13 +1,13 @@
 import Header from "../components/Header";
 import { Article } from "../shared/GlobalStyled";
-import ExhibitionList from "../features/exhibition/exhibitoinList/ExhibitionList";
+import ExhibitionLists from "../features/exhibition/exhibitoinList/ExhibitionLists";
 
 function Exhibition() {
   return (
     <>
       <Header />
       <Article>
-        <ExhibitionList />
+        <ExhibitionLists />
       </Article>
     </>
   );
